requirements.txt
jupyter
pin
example-robot-data
meshcat
ipywidgets
ipympl
roboticstoolbox-python
black
pycapacity
scipy<=1.11 # remove after roboticstoolbox-python fixes #412
